refactor(gitlab-ce-docker): move exec.go to internal/pkg/plugin
